Wrapped ReSID as sid_wrapper.
Basicly the idea would be that we generate n samples, then set parameters
and generate m samples more. Of course the input is now those parameters
which are also exposed outside. This is wrong because LFO's and external
ADSR's also want to manipulate the parameters, but their own parameters
should be interpolated as well.

Probably there is need for an yet another abstraction layer (yuck),
which takes parameter changes from outside, manipulates those with LFO's
and envgens, and passes them to the sid_wrapper.

One really has to figure out which layer takes care of which
interpolation. In sid_wrapper it would probably make sense
to give each parameter their own update rate (in cycles). This would be
then used by both manual manipulation, and LFO/ADSR-induced manipulation.

But then again, this might be also useful for the upper layer (that runs
the lfo's/adsr's). Of course we are working at audio rate here, so it's
not directly applicable. Let's think about this a bit..